RISK Game
---RISK Game [Postgres -> (NoSQL) DynamoDB, MongoDB integration] ](https://github.com/javaknows/javaknows/blob/DB/RiskGame.zip)

[+4]The aim of this project is to develop a prototype for the databases system properties necessary to create a digital risk board game.
Risk is a strategy board game of diplomacy, conflict and conquest for two to six players. [For rules google RISK game rules]
Since NoSQL databases have grown in popularity over the years, the underlying idea is not new. The special underlying technique that provides new solutions that improve reliability and performance. Hence, the Postgres, DynamoDB, and MongoDB work collaterally building a secure commercial environment. Likewise, subsequent relational fixes were achieved in order to build a self-managed infrastructure. Moreover, there might be an increase of the audience & number of users. 
Nowadays, it is better to focus on creating a combination of relational databases and NoSQL; however, we should not get used to a preferred solution for the simple fact of basing ourselves on subjective reasons, but rather allow new technologies that offer us new and best tools for a specific scheme.